198 Include user's card labels when generating the report (#219)
Closes #198 and accompanied by
https://github.com/insightsengineering/teal.modules.clinical/pull/835

# The problem

Currently most of the `teal` modules, that wrap up `teal.reporter` in
their UI, have titles of the `ReportCard` hardcoded in their body. For
example check out below 2 modules


https://github.com/insightsengineering/teal.modules.clinical/blob/2b20a85b22f25f15f6808f745b4d1d5333b123f3/R/tm_g_ci.R#L484-L503


https://github.com/insightsengineering/teal.modules.clinical/blob/2b20a85b22f25f15f6808f745b4d1d5333b123f3/R/tm_g_km.R#L774-L793

Such hardcoded content in `card$set_name()` and `card$append_text()` has
a great impact on the final `ReportCard`.
If `label` in the `Add Card` module UI is not provided, then the card
name is set to the one passed in `card$set_name()`. This is fine for
cards that had empty labels - they will have a name populated by the
developer of the teal module.


![image](https://github.com/insightsengineering/teal.reporter/assets/133694481/b5d3656b-b61f-450f-ba69-b9db14889a24)

However `card$append_text()` like this

https://github.com/insightsengineering/teal.modules.clinical/blob/2b20a85b22f25f15f6808f745b4d1d5333b123f3/R/tm_g_km.R#L778
and this

https://github.com/insightsengineering/teal.modules.clinical/blob/2b20a85b22f25f15f6808f745b4d1d5333b123f3/R/tm_g_ci.R#L489
appends a content to the card. The card gets a hardcoded title, no
matter what user specifies as a card label.

# The solution

`add_card_button_srv` allows to specify `card_fun` with `label`
parameter for card's title & content customization. This means, that you
are able to use `label` shiny input and use it in your `card_fun` to
pass the title. We are able to rewrite teal modules, so that they have
default titles, if `label` is empty in `Add Card` module UI. When it is
not empty, then the `label` is used as a name and as a title of the card
- like in this commit
https://github.com/insightsengineering/teal.modules.clinical/commit/5833c6c7f52aecac8da09b16da31e73fe5dffe07

## With label


![image](https://github.com/insightsengineering/teal.reporter/assets/133694481/2c96ba72-19eb-41c8-82a6-bd89bcaa02cc)

![image](https://github.com/insightsengineering/teal.reporter/assets/133694481/5ab17ab9-e2a9-40fc-96ee-068be0c34b42)


## Without the label


![image](https://github.com/insightsengineering/teal.reporter/assets/133694481/c318d90b-4518-4839-bf5a-5189e771c6bb)
**so the default `teal` module name and title are used**

![image](https://github.com/insightsengineering/teal.reporter/assets/133694481/7c313c1c-adf6-4cdd-afba-d23791144dbd)

---------

Signed-off-by: Marcin <133694481+m7pr@users.noreply.github.com>
Co-authored-by: kartikeya kirar <kirar.kartikeya1@gmail.com>
Co-authored-by: 27856297+dependabot-preview[bot]@users.noreply.github.com <27856297+dependabot-preview[bot]@users.noreply.github.com>
Co-authored-by: Dony Unardi <donyunardi@gmail.com>