Add a new regexp to the r-cmd-check-note blocklist (#291)
This PR is to add the notes to the CI part per
https://github.com/insightsengineering/teal.reporter/issues/288 but only
for teal.reporter.

The changes do not look to the general issue as vignettes failing to
build could be due to a number of reasons. As far as I know the string I
added is specific for this error. To fix the issue PR #290 was already
merge so I don't expect any failures now.

This could be added to the default/general check from r.pkg.template if
it meets the IDR requirements.

---------

Signed-off-by: Lluís Revilla <185338939+llrs-roche@users.noreply.github.com>
Co-authored-by: Marcin <133694481+m7pr@users.noreply.github.com>