436 Allow for larger file upload to upload large reports (#437)
# Pull Request

Fixes #436 

Added a package option that will control the size of the `fileInput`,
which [by default is 5
MB](https://mastering-shiny.org/action-transfer.html#server).
The option is a new `teal.reporter` option and by default is 10MB. With
that modification we allow:

1. Control for developers to decide if they allow big reports to be
uploaded.
2. Unblock the possibility of having reports larger than 5MB.

## Questions

* This implementation will modify all allowed max sizes of shiny
fileInputs. Is this a problem?
* Do I need to document this option? I do not see specific documentation
of other `teal.reporter` package options in vignettes or other parts.

---------

Signed-off-by: Marcin <133694481+m7pr@users.noreply.github.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Marcin <133694481+m7pr@users.noreply.github.com>